requirements.txt
python-docx>=1.1.0
openrouter>=0.3.0
pydantic>=2.0
pywin32>=306
Pillow>=10.0
lxml>=5.0
python-dotenv>=1.0
